# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The Ammonite API interface.",
        "title": "Ammonite API",
        "contact": {},
        "version": "0.1"
    },
    "basePath": "/api/",
    "paths": {
        "/decks": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deck"
                ],
                "summary": "Get all decks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Deck"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/decks/": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deck"
                ],
                "summary": "Create one deck",
                "parameters": [
                    {
                        "description": "Deck",
                        "name": "deck",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.InputDeck"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/decks/calibrate/{name}/{x}/{y}/{z}/{qw}/{qx}/{qy}/{qz}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deck"
                ],
                "summary": "Calibrate a deck",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Deck name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "X coordinate",
                        "name": "x",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Y coordinate",
                        "name": "y",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Z coordinate",
                        "name": "z",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Qw coordinate",
                        "name": "qw",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Qx coordinate",
                        "name": "qx",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Qy coordinate",
                        "name": "qy",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Qz coordinate",
                        "name": "qz",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/decks/{name}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deck"
                ],
                "summary": "Get one deck",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Deck name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Deck"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deck"
                ],
                "summary": "Delete one deck",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Deck name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/labwares": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "labware"
                ],
                "summary": "Get all labwares",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Labware"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/labwares/": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "labware"
                ],
                "summary": "Create one labware",
                "parameters": [
                    {
                        "description": "Labware",
                        "name": "labware",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Labware"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/labwares/{name}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "labware"
                ],
                "summary": "Get one labware",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Labware name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Labware"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "labware"
                ],
                "summary": "Delete one labware",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Labware name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "dev"
                ],
                "summary": "A pingable endpoint",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Message"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Deck": {
            "type": "object",
            "properties": {
                "calibrated": {
                    "type": "boolean"
                },
                "locations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.Location"
                    }
                },
                "name": {
                    "type": "string"
                },
                "qw": {
                    "type": "number"
                },
                "qx": {
                    "type": "number"
                },
                "qy": {
                    "type": "number"
                },
                "qz": {
                    "type": "number"
                },
                "x": {
                    "type": "number"
                },
                "y": {
                    "type": "number"
                },
                "z": {
                    "type": "number"
                }
            }
        },
        "main.InputDeck": {
            "type": "object",
            "properties": {
                "locations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.Location"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "main.Labware": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "wells": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.Well"
                    }
                },
                "zDimension": {
                    "description": "XDimension float64 `json:\"xDimension\"` // Microplate: 127.76\nYDimension float64 `json:\"yDimension\"` // Microplate: 85.48",
                    "type": "number"
                }
            }
        },
        "main.Location": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "x": {
                    "type": "number"
                },
                "y": {
                    "type": "number"
                },
                "z": {
                    "type": "number"
                }
            }
        },
        "main.Message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "main.Well": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "depth": {
                    "type": "number"
                },
                "diameter": {
                    "type": "number"
                },
                "x": {
                    "type": "number"
                },
                "y": {
                    "type": "number"
                },
                "z": {
                    "type": "number"
                }
            }
        }
    }
}